// headless-cms-api/src/common/dto/query.dto.ts
import { z } from 'zod';

export const QuerySchema = z.object({
  page: z.coerce.number().int().min(1).optional().default(1),
  limit: z.coerce.number().int().min(1).max(100).optional().default(10),
  sortBy: z.string().optional(),
  sortOrder: z.enum(['asc', 'desc']).optional().default('asc'),
  search: z.string().optional(),
});

export type QueryDto = z.infer<typeof QuerySchema>;